+++ src/main/java/me/discordlinking/Main.java
@@ -2,6 +2,10 @@ package me.discordlinking;
 
 import club.minnced.discord.webhook.WebhookClient;
 import club.minnced.discord.webhook.send.WebhookMessageBuilder;
+import me.discordlinking.commands.DisableBotCommand;
+import me.discordlinking.commands.DiscordCommand;
+import me.discordlinking.commands.EnableBotCommand;
+import me.discordlinking.commands.ReloadCommand;
 import net.dv8tion.jda.api.entities.Activity;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.java.JavaPlugin;
